+++ src/socket-as-promised.ts
@@ -170,7 +170,7 @@ export class SocketAsPromised implements AsyncIterable<IncomingPacket> {
     })
   }
 
-  address(): AddressInfo | string {
+  address(): AddressInfo {
     return this.socket.address()
   }